Task/remove old ir versions (#374)
* wip: remove ir-versions

* wip: remove 202209 and some 202304 files

* wip: add new fields to types and elasticsearch

* task: remove old ir versions
