+++ apps/innovations/_services/innovation-document.service.ts
@@ -3,12 +3,8 @@ import { injectable } from 'inversify';
 import type { EntityManager } from 'typeorm';
 
 import { ServiceRoleEnum } from '@innovations/shared/enums';
-import { InnovationErrorsEnum, NotFoundError } from '@innovations/shared/errors';
-import {
-  CurrentDocumentConfig,
-  DocumentType,
-  DocumentTypeFromVersion
-} from '@innovations/shared/schemas/innovation-record';
+import { ConflictError, InnovationErrorsEnum, NotFoundError } from '@innovations/shared/errors';
+import type { DocumentType } from '@innovations/shared/schemas/innovation-record';
 
 import { BaseService } from './base.service';
 import { InnovationDocumentDraftEntity, InnovationDocumentEntity } from '@innovations/shared/entities';
@@ -21,45 +17,42 @@ export class InnovationDocumentService extends BaseService {
   }
 
   /**
-   * TODO: Change this when changing the Return of DocumentTypeFromVersion
    * retrieves the latest version of a document from the database and validates the version
    * @param innovationId the innovation id
-   * @param version version of the document to retrieve
-   * @param type submitted or draft type
+   * @param params an object containing the type and version of the document to retrieve
    * @param entityManager optional entity manager for running inside transaction
    * @returns the document
    */
-  async getInnovationDocument<V extends DocumentType['version'], T extends DocumentTypeFromVersion<V>>(
+  async getInnovationDocument(
     innovationId: string,
-    version: V,
-    type: 'SUBMITTED' | 'DRAFT',
+    params: { type: 'SUBMITTED' | 'DRAFT'; version?: number },
     entityManager?: EntityManager
-  ): Promise<T> {
+  ): Promise<DocumentType> {
     const connection = entityManager ?? this.sqlConnection.manager;
+    // NOTE: Currently this type is not correct if a old version is fetched (doesn't happen currently.)
     let document: DocumentType | undefined;
 
-    if (version === CurrentDocumentConfig.version) {
+    if (!params.version) {
       document = (
         await connection
           .createQueryBuilder(
-            type === 'SUBMITTED' ? InnovationDocumentEntity : InnovationDocumentDraftEntity,
+            params.type === 'SUBMITTED' ? InnovationDocumentEntity : InnovationDocumentDraftEntity,
             'document'
           )
           .where('document.id = :innovationId', { innovationId })
-          // .andWhere('document.version = :version', { version })
           .getOne()
       )?.document;
     } else {
       const raw = await connection.query(
         `
         SELECT TOP 1 document
-        FROM ${type === 'SUBMITTED' ? 'innovation_document' : 'innovation_document_draft'}
+        FROM ${params.type === 'SUBMITTED' ? 'innovation_document' : 'innovation_document_draft'}
         FOR SYSTEM_TIME ALL
         WHERE id = @0
         AND version = @1
         ORDER BY valid_to DESC
       `,
-        [innovationId, version]
+        [innovationId, params.version]
       );
 
       document = raw.length ? JSON.parse(raw[0].document) : undefined;
@@ -69,11 +62,11 @@ export class InnovationDocumentService extends BaseService {
       throw new NotFoundError(InnovationErrorsEnum.INNOVATION_NOT_FOUND);
     }
 
-    // if (document.version !== version) {
-    //   throw new ConflictError(InnovationErrorsEnum.INNOVATION_DOCUMENT_VERSION_MISMATCH);
-    // }
+    if (params.version && params.version !== document.version) {
+      throw new ConflictError(InnovationErrorsEnum.INNOVATION_DOCUMENT_VERSION_MISMATCH);
+    }
 
-    return document as T;
+    return document;
   }
 
   /**
@@ -87,12 +80,7 @@ export class InnovationDocumentService extends BaseService {
     transaction: EntityManager,
     params?: { updatedAt?: Date; description?: string }
   ): Promise<void> {
-    const draftDocument = await this.getInnovationDocument(
-      innovationId,
-      CurrentDocumentConfig.version,
-      'DRAFT',
-      transaction
-    );
+    const draftDocument = await this.getInnovationDocument(innovationId, { type: 'DRAFT' }, transaction);
     await transaction.update(
       InnovationDocumentEntity,
       { id: innovationId },
