+++ libs/shared/schemas/innovation-record/document.types.ts
@@ -1,37 +1,54 @@
+import type {
+  InnovationSections,
+  catalogEvidenceSubmitType,
+  catalogEvidenceType,
+  catalogCategory,
+  catalogAreas,
+  catalogCareSettings,
+  catalogInvolvedAACProgrammes,
+  catalogKeyHealthInequalities,
+  catalogNeedsSupportAnyArea,
+  catalogStandardsType
+} from './catalog.types';
+
 type OtherKeyValues = { [key: string]: any };
 
 export type InnovationRecordDocumentType = {
   version: number;
   INNOVATION_DESCRIPTION: {
-    name: string; // N
-    description?: string; // N
-    postcode?: string; // N
-    countryName?: string; // N
-    categories?: string[]; // N
-    otherCategoryDescription?: string; // N
-    mainCategory?: string; // N
-    areas?: string[]; // N
-    careSettings?: string[]; // N
-    otherCareSetting?: string; // N
-    involvedAACProgrammes?: string[]; // N
+    name: string;
+    description?: string;
+    postcode?: string;
+    countryName?: string;
+    officeLocation?: string;
+    countryLocation?: string;
+    categories?: catalogCategory[];
+    otherCategoryDescription?: string;
+    mainCategory?: catalogCategory;
+    areas?: catalogAreas[];
+    careSettings?: catalogCareSettings[];
+    otherCareSetting?: string;
+    involvedAACProgrammes?: catalogInvolvedAACProgrammes[];
+    hasWebsite?: string;
+    website?: string;
   } & OtherKeyValues;
   UNDERSTANDING_OF_NEEDS: {
-    diseasesConditionsImpact?: string[]; // N
-    keyHealthInequalities?: string[]; // N
+    diseasesConditionsImpact?: string[];
+    keyHealthInequalities?: catalogKeyHealthInequalities[];
   } & OtherKeyValues;
   EVIDENCE_OF_EFFECTIVENESS: {
     hasEvidence?: string;
     currentlyCollectingEvidence?: string;
     summaryOngoingEvidenceGathering?: string;
-    needsSupportAnyArea?: string[];
+    needsSupportAnyArea?: catalogNeedsSupportAnyArea[];
   } & OtherKeyValues;
   MARKET_RESEARCH: OtherKeyValues;
   CURRENT_CARE_PATHWAY: OtherKeyValues;
   TESTING_WITH_USERS: {
-    userTests?: { kind: string; feedback?: string }[]; // N
+    userTests?: { kind: string; feedback?: string }[];
   } & OtherKeyValues;
   REGULATIONS_AND_STANDARDS: {
-    standards?: { type: string; hasMet?: string }[]; // N
+    standards?: { type: catalogStandardsType; hasMet?: string }[];
   } & OtherKeyValues;
   INTELLECTUAL_PROPERTY: OtherKeyValues;
   REVENUE_MODEL: OtherKeyValues;
@@ -39,29 +56,32 @@ export type InnovationRecordDocumentType = {
   DEPLOYMENT: OtherKeyValues;
   evidences?: {
     id: string;
-    evidenceSubmitType: string;
-    evidenceType?: string;
+    evidenceSubmitType: catalogEvidenceSubmitType;
+    evidenceType?: catalogEvidenceType;
     description?: string;
     summary: string;
   }[];
 };
 
 // Required Sections/Questions
-export const requiredSectionsAndQuestions = new Map([
+export const requiredSectionsAndQuestions = new Map<InnovationSections, string[]>([
   [
     'INNOVATION_DESCRIPTION',
     [
       'name',
       'description',
       'postcode',
-      // 'countryName', // TODO: Enable this after doing the calculatedFields
+      'officeLocation',
+      'countryLocation',
       'categories',
       'otherCategoryDescription',
       'mainCategory',
       'areas',
       'careSettings',
       'otherCareSetting',
-      'involvedAACProgrammes'
+      'involvedAACProgrammes',
+      'hasWebsite',
+      'website'
     ]
   ],
   ['UNDERSTANDING_OF_NEEDS', ['diseasesConditionsImpact', 'keyHealthInequalities']],
@@ -72,12 +92,7 @@ export const requiredSectionsAndQuestions = new Map([
   ['MARKET_RESEARCH', []],
   ['CURRENT_CARE_PATHWAY', []],
   ['TESTING_WITH_USERS', ['userTests']],
-  [
-    'REGULATIONS_AND_STANDARDS',
-    [
-      // 'standards' // TODO: Enable this after solving the fields-groups
-    ]
-  ],
+  ['REGULATIONS_AND_STANDARDS', ['standards']],
   ['INTELLECTUAL_PROPERTY', []],
   ['REVENUE_MODEL', []],
   ['COST_OF_INNOVATION', []],
